+++ tasks-jsf/src/test/java/org/jboss/as/quickstarts/tasksJsf/TaskListBeanTest.java
@@ -19,10 +19,7 @@ package org.jboss.as.quickstarts.tasksJsf;
 import static org.junit.Assert.assertEquals;
 
 import java.io.FileNotFoundException;
-import java.util.Arrays;
-import java.util.List;
 
-import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
 
 import org.jboss.arquillian.container.test.api.Deployment;
@@ -46,12 +43,12 @@ public class TaskListBeanTest {
 
     @Deployment
     public static WebArchive deployment() throws IllegalArgumentException, FileNotFoundException {
-        return new DefaultDeployment().withPersistence().withImportedData().getArchive()
-                .addClasses(User.class, Task.class, TaskList.class, TaskListBean.class, TaskDao.class);
+        return new DefaultDeployment(true).withPersistence().withImportedData().getArchive()
+                .addClasses(User.class, Task.class, TaskList.class, TaskListBean.class, TaskDao.class, TaskDaoStub.class, Testing.class);
     }
 
     @Inject
-    private TaskDaoStub taskDaoStub;
+    private TaskDao taskDaoStub;
 
     @Inject
     private TaskList taskList;
@@ -61,52 +58,18 @@ public class TaskListBeanTest {
         taskList.getAll();
         taskList.getAll();
         taskList.getAll();
-        assertEquals(1, taskDaoStub.getGetAllCallsCount());
+        assertEquals(1, ((TaskDaoStub) taskDaoStub).getGetAllCallsCount());
     }
 
     @Test
     public void dao_method_getAll_should_be_called_after_invalidation() {
         taskList.getAll();
         taskList.getAll();
-        assertEquals(1, taskDaoStub.getGetAllCallsCount());
+        assertEquals(1, ((TaskDaoStub) taskDaoStub).getGetAllCallsCount());
         taskList.invalidate();
-        assertEquals(1, taskDaoStub.getGetAllCallsCount());
+        assertEquals(1, ((TaskDaoStub) taskDaoStub).getGetAllCallsCount());
         taskList.getAll();
         taskList.getAll();
-        assertEquals(2, taskDaoStub.getGetAllCallsCount());
-    }
-
-    @RequestScoped
-    public static class TaskDaoStub implements TaskDao {
-
-        private int getAllCallsCount = 0;
-
-        @Override
-        public void createTask(User user, Task task) {
-        }
-
-        @Override
-        public List<Task> getAll(User user) {
-            getAllCallsCount += 1;
-            return Arrays.asList(new Task[] {});
-        }
-
-        @Override
-        public List<Task> getRange(User user, int offset, int count) {
-            return null;
-        }
-
-        @Override
-        public List<Task> getForTitle(User user, String title) {
-            return null;
-        }
-
-        @Override
-        public void deleteTask(Task task) {
-        }
-
-        public int getGetAllCallsCount() {
-            return getAllCallsCount;
-        }
+        assertEquals(2, ((TaskDaoStub) taskDaoStub).getGetAllCallsCount());
     }
 }
